graph-simplification: play with new ideas for merge-borders
I think it might be a good idea to prevent collapsing nodes from different
blocks together.
While it makes the graph look simpler, it will make navigation _very_ weird,
as some nodes in the simplified graph may extend way beyond what the user expect.
This will make it unintuitive for code navigation.

So the current idea is to add a `block` identifier (similar to clusters)
and prevent node-merges across `block` boundaries.

It seems that fully-automating this won't work (a block is one thing,
a `switch` or `match` or similar is a different thing for the AST),
but I might be able to do it decleratively with a list of "block-defining"
AST nodes.